Move controller from wpilibj to wpimath
