+++ wpimath/src/main/java/edu/wpi/first/math/controller/PIDController.java
@@ -2,10 +2,10 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package edu.wpi.first.wpilibj.controller;
+package edu.wpi.first.math.controller;
 
-import edu.wpi.first.hal.FRCNetComm.tResourceType;
-import edu.wpi.first.hal.HAL;
+import edu.wpi.first.math.MathSharedStore;
+import edu.wpi.first.math.MathUsageId;
 import edu.wpi.first.math.MathUtil;
 import edu.wpi.first.util.sendable.Sendable;
 import edu.wpi.first.util.sendable.SendableBuilder;
@@ -88,7 +88,7 @@ public class PIDController implements Sendable, AutoCloseable {
     instances++;
     SendableRegistry.addLW(this, "PIDController", instances);
 
-    HAL.report(tResourceType.kResourceType_PIDController2, instances);
+    MathSharedStore.reportUsage(MathUsageId.kController_PIDController2, instances);
   }
 
   @Override
